move `pkg/api` to `pkg/admin/api`.
