+++ pkg/admin/api/cfg.yaml
