--- pkg/api/.gitattributes
@@ -1,2 +1,0 @@
-*.js filter=lfs diff=lfs merge=lfs -text
-*.css filter=lfs diff=lfs merge=lfs -text
